// src/types/alarm.ts
export interface AlarmListItem {
  alarmNo: number
  title: string
  content: string
}

export interface CompanyAlarmApproveResponse {
  CompanyName: string
  affiliationName: string
  branchName: string
  phoneNumber: string
  address: string
  createdAt: string
  updatedAt: string
  message: string
}